// src/models/enums/appointment-status.ts
export enum AppointmentStatus {
    PENDING_USER_CONFIRMATION = "PENDING_USER_CONFIRMATION",
    PENDING_ORGANIZATION_APPROVAL = "PENDING_ORGANIZATION_APPROVAL",
    CONFIRMED = "CONFIRMED",
    REJECTED = "REJECTED",
    CANCELLED = "CANCELLED",
    IN_PROGRESS = "IN_PROGRESS",
    COMPLETED = "COMPLETED",
    NO_SHOW = "NO_SHOW",
}